sway: use plsdo wherever possible
On sway, we want to reset the keybinding layer before calling certain
commands. My first though was to include this into plsdo, so I exported
an env variable in the shell profile that plsdo can read to determine
which window manager / compositor is running, to know what command to
invoke (swaymsg / rivertctl / etc...). This wasn't needed in the end but
I left it in - it might be needed later and setting an env var manually
seems to be the cleanest way, and I might forget