eww: avoid onclick command timeout
Apparently there is a 200ms timeout for commands executed by onclick and
such events, and this timeout cannot be overridden. However, my keyboard
layout switcheroo command pops up a dmenu that must wait for user input.
The dmenu window stays functional after the 200ms passes, but the rest
of the command does not run, so nothing happens after I choose a layout.

It seems the timeout can be avoided by putting the command in the
background with `&`

https://github.com/elkowar/eww/issues/626